backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
supabase==2.9.1
pydantic==2.9.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
httpx==0.27.0
apscheduler==3.10.4
python-multipart==0.0.9
reportlab==4.2.0
openpyxl==3.1.3
google-generativeai==0.8.0
python-dotenv==1.0.1
